Move SDK modules factory into core
