+++ src/app/core/factory/sdk/index.ts
@@ -4,9 +4,9 @@ import {
   ApiSecurity, ApiUrl,
   AppDetails
 } from '@internxt/sdk/dist/shared';
-import packageJson from '../../../package.json';
-import localStorageService from '../../app/core/services/local-storage.service';
-import { LocalStorageItem, Workspace } from '../../app/core/types';
+import packageJson from '../../../../../package.json';
+import localStorageService from '../../services/local-storage.service';
+import { LocalStorageItem, Workspace } from '../../types';
 
 export function createAuthClient(): Auth {
   const apiUrl = getApiUrl();
