+++ src/app/core/views/DeactivationView/DeactivationView.tsx
@@ -13,7 +13,7 @@ import notificationsService, { ToastType } from '../../../notifications/services
 import { match } from 'react-router-dom';
 import navigationService from '../../services/navigation.service';
 import { AppView } from '../../types';
-import { createAuthClient } from '../../../../factory/modules';
+import { createAuthClient } from '../../factory/sdk';
 
 export interface DeactivationViewProps {
   match?: match<{ token: string }>;
