+++ src/app/drive/services/file.service/uploadFile.ts
@@ -9,7 +9,7 @@ import navigationService from '../../../core/services/navigation.service';
 import { getEnvironmentConfig, Network } from '../network';
 import { encryptFilename } from '../../../crypto/services/utils';
 import errorService from '../../../core/services/error.service';
-import { createStorageClient } from '../../../../factory/modules';
+import { createStorageClient } from '../../../core/factory/sdk';
 
 export interface ItemToUpload {
   name: string;
