+++ src/app/payment/services/products.service.ts
@@ -1,5 +1,5 @@
 import { ProductData } from '../types';
-import { createPaymentsClient } from '../../../factory/modules';
+import { createPaymentsClient } from '../../core/factory/sdk';
 
 export const fetchProducts = async (): Promise<ProductData[]> => {
   return createPaymentsClient().getProducts();
