+++ src/app/share/services/share.service.ts
@@ -1,5 +1,5 @@
 import errorService from '../../core/services/error.service';
-import { createShareClient } from '../../../factory/modules';
+import { createShareClient } from '../../core/factory/sdk';
 import { ShareTypes } from '@internxt/sdk/dist/drive';
 
 export function generateShareLink(params: ShareTypes.GenerateShareLinkPayload): Promise<string> {
